organized functions into modular files - cleaner index.js
