+++ src/functions/getRandomNumFromLength.js
@@ -1,0 +1,7 @@
+function getRandomNumFromLength(size) {
+    return Math.floor(Math.random() * size);
+}
+
+module.exports = {
+    getRandomNumFromLength
+};
